refactor: authorization for owner on task controller

+++ src/tasks/tasks.service.ts
@@ -14,12 +14,15 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) {}
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(
+        filterDto: GetTasksFilterDto, 
+        user: User
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne({ where: { id }} );
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id }} );
 
         if(!found) {
             throw new NotFoundException(`Task with id "${id}" not found`); 
@@ -28,20 +31,27 @@ export class TasksService {
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-       return await this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+       return await this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete({ id });  
+    async deleteTask(
+        id: number, 
+        user: User
+    ): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });  //different between delete and findOne --- where
 
         if(result.affected === 0) {
             throw new NotFoundException(`Task with id "${id}" not found`);
         }       
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus) : Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus, 
+        user: User,
+    ) : Promise<Task> {
+        const task = await this.getTaskById(id, user);
 
         task.status = status;
         await task.save();
